major changes
couldn't get the program running in the old project after the changes i
made, so i created a new repo in a new project and copied our stuff into
it. since then i've made some changes. i like the 45-deg angle thing but
there's still a problem w turning. it can follow the wall just fine
until it turns, esp w a left turn. i've tried playing around with the
speeds and distances and have gotten it to work w right turns most of
the time, but with left turns it goes too wide or just crashes into the
wall since it can't detect how close it is getting.

ignore the commented out stuff with timers and whatever - i was trying
to see what would happen if i tried to delay for gaps and large left
turns (ie the distance is bigger than normal). basically just go forward
for 4s and if it's still a high distance after that turn sharp left.
otherwise let the problem correct itself for issues w the sensor or gaps
between the blocks.

we may need to play with the directions the wheels turn in, or even
stopping them to control the sharpness of our turns. i don't think this
will work just by swerving. the distances are too small compared to our
robot. message me w any thoughts.